% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs-plot.R
\name{plot_cumulative_majority}
\alias{plot_cumulative_majority}
\title{Plot cumulative issue majority vs individual issue majority by group}
\usage{
plot_cumulative_majority(
  results,
  show_response_value = TRUE,
  majority_cutoff = FALSE
)
}
\arguments{
\item{results}{Output from \code{measure_alignment()}.}

\item{show_response_value}{Logical; show response values in labels.}

\item{majority_cutoff}{Logical; trim items that fall below majority cumulative support in each group.}
}
\value{
ggplot object.
}
\description{
For each group, plot support for items in descending order of individual
issue majority, overlaying the individual issue majority vs. the cumulative
issue majority (proportion supporting the majority on all issues up to and
including rank k).
}
