% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs-plot.R
\name{plot_cumulative_support}
\alias{plot_cumulative_support}
\title{Plot cumulative issue support by group}
\usage{
plot_cumulative_support(
  results,
  show_response_value = TRUE,
  majority_cutoff = FALSE
)
}
\arguments{
\item{results}{Output from \code{measure_alignment()}.}

\item{show_response_value}{Logical; show response values in labels.}

\item{majority_cutoff}{Logical; trim items that fall below majority cumulative support in each group.}
}
\value{
ggplot object.
}
\description{
For each group, plot support for items in descending order,
overlaying single item support vs. cumulative support.
}
