% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs-plot.R
\name{plot_group_alignment}
\alias{plot_group_alignment}
\title{Plot alignment distributions and summary means by group.}
\usage{
plot_group_alignment(
  results,
  exclude_vals = "DK/REF",
  type = c("density", "histogram"),
  show_distributions = FALSE,
  show_means = TRUE,
  ci_scale = 1.96
)
}
\arguments{
\item{results}{Output from \code{measure_alignment()}.}

\item{exclude_vals}{Group values to exclude.}

\item{type}{Density or histogram.}

\item{show_distributions}{Logical; show distribution shapes.}

\item{show_means}{Logical; show mean estimates and intervals.}

\item{ci_scale}{Multiplier for confidence intervals.}
}
\value{
ggplot object.
}
\description{
Plot alignment distributions and summary means by group.
}
