% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs-plot.R
\name{plot_individual_alignments}
\alias{plot_individual_alignments}
\title{Plot alignment distributions and summary means by group.}
\usage{
plot_individual_alignments(
  results,
  exclude_vals = "DK/REF",
  show_dists_as = c("histogram", "density", "none"),
  show_mean = TRUE,
  ci_scale = 1.96
)
}
\arguments{
\item{results}{Output from \code{measure_alignment()}.}

\item{exclude_vals}{Group values to exclude.}

\item{show_dists_as}{How to visualize distributions.}

\item{show_mean}{Logical; show mean estimates and intervals.}

\item{ci_scale}{Multiplier for confidence intervals.}
}
\value{
ggplot object.
}
\description{
Plot alignment distributions and summary means by group.
}
