% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs-align.R
\name{print.survalign}
\alias{print.survalign}
\title{Print method for survalign objects.}
\usage{
\method{print}{survalign}(x, ...)
}
\arguments{
\item{x}{An object of class \code{'survalign'}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for survalign objects.
}
