% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survalign-package.R
\docType{package}
\name{survalign-package}
\alias{survalign}
\alias{survalign-package}
\title{survalign: Measure Group Alignment in Survey Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for measuring public opinion alignment within groups using cross-sectional or time series survey data.
}
\author{
\strong{Maintainer}: Soubhik Barari \email{barari-soubhik@norc.org}

}
\keyword{internal}
