% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted_se}
\alias{weighted_se}
\title{Design-weighted standard error of the mean (internal)}
\usage{
weighted_se(x, w)
}
\description{
Uses the Kish effective sample size approximation:
n_eff = (sum w)^2 / sum(w^2)
so SE = sd_w / sqrt(n_eff), where sd_w is the weighted standard deviation.
}
