% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dump_labels}
\alias{dump_labels}
\title{Dump column labels to dataframe (internal)}
\usage{
dump_labels(data, attr = c("label", "shortlabel"), file = NULL)
}
\description{
Dump column labels to dataframe (internal)
}
