% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs-plot.R
\name{plot_alignment_curve}
\alias{plot_alignment_curve}
\title{Plot cumulative alignment curves by group.}
\usage{
plot_alignment_curve(
  results,
  group_label = attr(results, "group_label"),
  group_colors = NULL,
  exclude_vals = "DK/REF",
  resolution = seq(0, 1, by = 0.01)
)
}
\arguments{
\item{results}{Output from \code{analyze_opinion_coalitions}.}

\item{group_label}{Optional label for legend.}

\item{group_colors}{Optional manual colors.}

\item{exclude_vals}{Group values to exclude.}

\item{resolution}{Vector of quantiles to evaluate.}
}
\value{
ggplot object.
}
\description{
Plot cumulative alignment curves by group.
}
