% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_ques}
\alias{resolve_ques}
\title{Resolve columns corresponding to questions (internal)}
\usage{
resolve_ques(data, ques_cols = NULL, ques_stem = NULL, exclude_cols = NULL)
}
\description{
Resolve columns corresponding to questions (internal)
}
